+++ packages/press/src/entry.ssr.tsx
@@ -4,23 +4,22 @@ import {
   unstable_routeRSCServerRequest as routeRSCServerRequest,
   unstable_RSCStaticRouter as RSCStaticRouter,
 } from 'react-router';
+import type { ReactFormState } from 'react-dom/client';
 
 export async function generateHTML(
   request: Request,
-  fetchServer: (request: Request) => Promise<Response>,
+  serverResponse: Response,
 ): Promise<Response> {
   return await routeRSCServerRequest({
     // The incoming request.
     request,
-    // How to call the React Server.
-    fetchServer,
+    // The React Server response
+    serverResponse,
     // Provide the React Server touchpoints.
     createFromReadableStream,
     // Render the router to HTML.
     async renderHTML(getPayload) {
       const payload = await getPayload();
-      const formState =
-        payload.type === 'render' ? await payload.formState : undefined;
 
       const bootstrapScriptContent =
         await import.meta.viteRsc.loadBootstrapScriptContent('index');
@@ -29,8 +28,10 @@ export async function generateHTML(
         <RSCStaticRouter getPayload={getPayload} />,
         {
           bootstrapScriptContent,
-          // @ts-expect-error - no types for this yet
-          formState,
+          formState:
+            'formState' in payload
+              ? (payload.formState as ReactFormState)
+              : undefined,
         },
       );
     },